refactor(state): 공통 기능 분리
