--- src/main/java/blackjack/domain/CardAddible.java
@@ -1,12 +1,0 @@
-package blackjack.domain;
-
-public interface CardAddible {
-
-    Cards cards();
-
-    void addCard(Card card);
-
-    boolean isEnd();
-
-    CardAddible judge();
-}
