+++ src/main/java/blackjack/domain/state/Gameable.java
@@ -1,0 +1,15 @@
+package blackjack.domain.state;
+
+import blackjack.domain.card.Card;
+import blackjack.domain.card.Cards;
+
+public interface Gameable {
+
+    Cards cards();
+
+    void addCard(Card card);
+
+    boolean isEnd();
+
+    Gameable judge();
+}
