+++ src/main/java/blackjack/domain/state/StateSupport.java
@@ -1,0 +1,18 @@
+package blackjack.domain.state;
+
+import blackjack.domain.card.Cards;
+
+public abstract class StateSupport extends CardAddibleSupport {
+
+    boolean end;
+
+    public StateSupport(final Cards card, boolean end) {
+        super(card);
+        this.end = end;
+    }
+
+    @Override
+    public boolean isEnd() {
+        return end;
+    }
+}
